src/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
supabase>=2.4.0
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
tavily-python>=0.3.0
pypdf>=4.2.0
python-pptx>=0.6.23
tiktoken>=0.7.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic>=2.7.0
httpx>=0.27.0
